Rename Internet -> Socket
Socket is more accurate.
