+++ engine/src/main/resources/default.cfg
@@ -100,7 +100,7 @@
     "serverPort": 25777,
     "masterServer": "meta.terasology.org"
   },
-  "internet": {
+  "socket": {
     "tcpWhitelists": {}
   }
 }
